fix: Interface 분리 제거
[TODO]
인터페이스를 분리하는 이유 고민하고 학습한 뒤에
STEP 7,8,9 에서 다시 생각해보자
