--- src/main/java/roomescape/reservation/dao/JdbcTemplateReservationDao.java
@@ -1,56 +1,0 @@
-package roomescape.reservation.dao;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Component;
-import roomescape.reservation.domain.Reservation;
-
-import javax.sql.DataSource;
-import java.util.List;
-
-@Component
-public class JdbcTemplateReservationDao implements ReservationDao {
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert jdbcInsert;
-
-    public JdbcTemplateReservationDao(final JdbcTemplate jdbcTemplate, final DataSource dataSource) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.jdbcInsert = new SimpleJdbcInsert(dataSource)
-                .withTableName("reservation")
-                .usingGeneratedKeyColumns("id");
-    }
-
-    @Override
-    public Reservation insert(final Reservation reservation) {
-        SqlParameterSource param = new BeanPropertySqlParameterSource(reservation);
-        long id = jdbcInsert.executeAndReturnKey(param).longValue();
-
-        return new Reservation(id, reservation);
-    }
-
-    @Override
-    public List<Reservation> findAll() {
-        return jdbcTemplate.query("SELECT * FROM reservation", reservationRowMapper);
-    }
-
-    @Override
-    public int deleteById(final Long id) {
-        String sql = "DELETE FROM reservation WHERE id = (?)";
-        int updateCount = jdbcTemplate.update(sql, id);
-
-        return updateCount;
-    }
-
-    private final RowMapper<Reservation> reservationRowMapper = (rs, rowNum) -> {
-        Reservation reservation = new Reservation(
-                rs.getLong("id"),
-                rs.getString("name"),
-                rs.getString("date"),
-                rs.getString("time")
-        );
-        return reservation;
-    };
-}
