+++ src/main/java/roomescape/reservation/dao/ReservationDao.java
@@ -1,14 +1,53 @@
 package roomescape.reservation.dao;
 
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Component;
 import roomescape.reservation.domain.Reservation;
 
+import javax.sql.DataSource;
 import java.util.List;
 
-public interface ReservationDao {
+@Component
+public class ReservationDao {
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert jdbcInsert;
 
-    Reservation insert(final Reservation reservation);
+    public ReservationDao(final JdbcTemplate jdbcTemplate, final DataSource dataSource) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.jdbcInsert = new SimpleJdbcInsert(dataSource)
+                .withTableName("reservation")
+                .usingGeneratedKeyColumns("id");
+    }
 
-    List<Reservation> findAll();
+    public Reservation insert(final Reservation reservation) {
+        SqlParameterSource param = new BeanPropertySqlParameterSource(reservation);
+        long id = jdbcInsert.executeAndReturnKey(param).longValue();
 
-    int deleteById(final Long id);
+        return new Reservation(id, reservation);
+    }
+
+    public List<Reservation> findAll() {
+        return jdbcTemplate.query("SELECT * FROM reservation", reservationRowMapper);
+    }
+
+    public int deleteById(final Long id) {
+        String sql = "DELETE FROM reservation WHERE id = (?)";
+        int updateCount = jdbcTemplate.update(sql, id);
+
+        return updateCount;
+    }
+
+    private final RowMapper<Reservation> reservationRowMapper = (rs, rowNum) -> {
+        Reservation reservation = new Reservation(
+                rs.getLong("id"),
+                rs.getString("name"),
+                rs.getString("date"),
+                rs.getString("time")
+        );
+        return reservation;
+    };
 }
