+++ src/test/java/roomescape/reservation/service/reservation/ReservationServiceTest.java
@@ -8,13 +8,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.reservation.dao.JdbcTemplateReservationDao;
+import roomescape.reservation.dao.ReservationDao;
 import roomescape.reservation.domain.repository.ReservationRepository;
-import roomescape.reservation.domain.repository.ReservationRepositoryImpl;
 import roomescape.reservation.dto.ReservationRequestDto;
 import roomescape.reservation.dto.ReservationResponseDto;
 import roomescape.reservation.service.ReservationService;
-import roomescape.reservation.service.ReservationServiceImpl;
 
 import javax.sql.DataSource;
 
@@ -22,7 +20,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @JdbcTest
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class ReservationServiceImplTest {
+class ReservationServiceTest {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
@@ -34,8 +32,8 @@ class ReservationServiceImplTest {
 
     @BeforeEach
     void init() {
-        ReservationRepository reservationRepository = new ReservationRepositoryImpl(new JdbcTemplateReservationDao(jdbcTemplate, dataSource));
-        reservationService = new ReservationServiceImpl(reservationRepository);
+        ReservationRepository reservationRepository = new ReservationRepository(new ReservationDao(jdbcTemplate, dataSource));
+        reservationService = new ReservationService(reservationRepository);
     }
 
     @DisplayName("예약 정보 삽입 테스트")
